Move (plugin-proxy) components to their own files.
part of CURA-10851
