+++ include/plugins/pluginproxy.h
@@ -5,6 +5,9 @@
 #define PLUGINS_PLUGINPROXY_H
 
 #include "Application.h"
+#include "components/common.h"
+#include "components/broadcast.h"
+#include "components/modify.h"
 #include "cura/plugins/slots/broadcast/v0/broadcast.grpc.pb.h"
 #include "cura/plugins/slots/broadcast/v0/broadcast.pb.h"
 #include "cura/plugins/slots/handshake/v0/handshake.grpc.pb.h"
@@ -34,225 +37,6 @@
 
 namespace cura::plugins
 {
-
-namespace details
-{
-
-template<typename T>
-concept plugin_modifier = requires(T value)
-{
-    requires std::is_member_function_pointer_v<decltype(&T::PrepareAsyncCall)>;
-};
-
-template<typename T>
-concept not_plugin_modifier = ! plugin_modifier<T>;
-
-using slot_info_ptr = std::shared_ptr<slot_metadata>;
-using plugin_info_ptr = std::shared_ptr<std::optional<plugin_metadata>>;
-
-} // namespace details
-
-/**
- * @brief Prepares client_context for the remote call.
- *
- * Sets timeout for the call and adds metadata to context.
- *
- * @param client_context - Client context to prepare
- * @param timeout - Call timeout duration (optional, default = 500ms)
- */
-inline void prep_client_context(grpc::ClientContext& client_context, const slot_metadata& slot_info, const std::chrono::milliseconds& timeout = std::chrono::milliseconds(500))
-{
-    // Set time-out
-    client_context.set_deadline(std::chrono::system_clock::now() + timeout);
-
-    // Metadata
-    client_context.AddMetadata("cura-engine-uuid", slot_info.engine_uuid.data());
-    client_context.AddMetadata("cura-thread-id", fmt::format("{}", std::this_thread::get_id()));
-}
-
-template<class Stub, typename RequestTp, typename ResponseTp>
-class PluginProxyModifyComponent
-{
-public:
-    constexpr PluginProxyModifyComponent();
-    PluginProxyModifyComponent(details::slot_info_ptr slot_info, details::plugin_info_ptr plugin_info, std::shared_ptr<grpc::Channel> channel);
-    auto modify(auto&&... args);
-};
-
-template<details::not_plugin_modifier Stub, typename RequestTp, typename ResponseTp>
-class PluginProxyModifyComponent<Stub, RequestTp, ResponseTp>
-{
-public:
-    constexpr PluginProxyModifyComponent() = default;
-
-    PluginProxyModifyComponent(details::slot_info_ptr slot_info, details::plugin_info_ptr plugin_info, std::shared_ptr<grpc::Channel> channel)
-    {
-    }
-
-    auto modify(auto&&... args)
-    {
-        assert(false); // Modify on a stub which it isn't meant for (for example, broadcast), should not actually be called.
-        return 0;
-    }
-};
-
-template<details::plugin_modifier Stub, utils::grpc_convertable RequestTp, utils::grpc_convertable ResponseTp>
-class PluginProxyModifyComponent<Stub, RequestTp, ResponseTp>
-{
-    using value_type = typename ResponseTp::native_value_type;
-
-    using req_converter_type = RequestTp;
-    using rsp_converter_type = ResponseTp;
-    using rsp_msg_type = typename ResponseTp::value_type;
-
-    using modify_stub_t = Stub;
-
-public:
-    constexpr PluginProxyModifyComponent() = default;
-
-    PluginProxyModifyComponent(details::slot_info_ptr slot_info, details::plugin_info_ptr plugin_info, std::shared_ptr<grpc::Channel> channel)
-        : slot_info_{ slot_info }
-        , plugin_info_{ plugin_info }
-        , modify_stub_{ channel }
-    {
-    }
-
-    /**
-     * @brief Executes to plugin Modify operation.
-     *
-     * As part of this operation, a request is sent to the plugin
-     * and the returned response is processed.
-     *
-     * @tparam Args -  argument types for the plugin request
-     * @param args - arguments for the plugin request
-     * @return The converted response value from plugin.
-     *
-     * @throws std::runtime_error if communication with the plugin fails.
-     */
-    auto modify(auto&&... args)
-    {
-        agrpc::GrpcContext grpc_context;
-        value_type ret_value{};
-        grpc::Status status;
-
-        boost::asio::co_spawn(
-            grpc_context,
-            [this, &grpc_context, &status, &ret_value, &args...]()
-            {
-                return this->modifyCall(grpc_context, status, ret_value, std::forward<decltype(args)>(args)...);
-            },
-            boost::asio::detached);
-        grpc_context.run();
-
-        if (! status.ok()) // TODO: handle different kind of status codes
-        {
-            if (plugin_info_->has_value())
-            {
-                throw exceptions::RemoteException(*slot_info_, plugin_info_->value(), status.error_message());
-            }
-            throw exceptions::RemoteException(*slot_info_, status.error_message());
-        }
-        return ret_value;
-    }
-
-private:
-    /**
-     * @brief Executes the modifyCall operation with the plugin.
-     *
-     * Sends a request to the plugin and saves the response.
-     *
-     * @param grpc_context - The gRPC context to use for the call
-     * @param status - Status of the gRPC call which gets updated in this method
-     * @param ret_value - Reference to the value in which response to be stored
-     * @param args - Request arguments
-     * @return A boost::asio::awaitable<void> indicating completion of the operation
-     */
-    boost::asio::awaitable<void> modifyCall(agrpc::GrpcContext& grpc_context, grpc::Status& status, value_type& ret_value, auto&&... args)
-    {
-        using RPC = agrpc::RPC<&modify_stub_t::PrepareAsyncCall>;
-        grpc::ClientContext client_context{};
-        prep_client_context(client_context, *slot_info_);
-
-        // Construct request
-        auto request{ req_(std::forward<decltype(args)>(args)...) };
-
-        // Make unary request
-        rsp_msg_type response;
-        status = co_await RPC::request(grpc_context, modify_stub_, client_context, request, response, boost::asio::use_awaitable);
-        ret_value = rsp_(response);
-        co_return;
-    }
-
-    req_converter_type req_{}; ///< The Modify request converter object.
-    rsp_converter_type rsp_{}; ///< The Modify response converter object.
-
-    details::slot_info_ptr slot_info_;
-    details::plugin_info_ptr plugin_info_;
-    ranges::semiregular_box<modify_stub_t> modify_stub_; ///< The gRPC Modify stub for communication.
-};
-
-template<plugins::v0::SlotID S>  // NOTE: Leave slot here (templated) for the case where we have to specialize broadcast-channels by slot.
-class PluginProxyBroadcastComponent
-{
-    using broadcast_stub_t = slots::broadcast::v0::BroadcastService::Stub;
-
-public:
-    constexpr PluginProxyBroadcastComponent() = default;
-
-    PluginProxyBroadcastComponent(details::slot_info_ptr slot_info, details::plugin_info_ptr plugin_info, std::shared_ptr<grpc::Channel> channel)
-        : slot_info_{ slot_info }
-        , plugin_info_{ plugin_info }
-        , broadcast_stub_{ channel }
-    {
-    }
-
-    template<plugins::v0::SlotID Subscription>
-    void broadcast(auto&&... args)
-    {
-        if (! plugin_info_->value().broadcast_subscriptions.contains(Subscription))
-        {
-            return;
-        }
-        agrpc::GrpcContext grpc_context;
-        grpc::Status status;
-
-        boost::asio::co_spawn(
-            grpc_context,
-            [this, &grpc_context, &status, &args...]()
-            {
-                return this->broadcastCall<Subscription>(grpc_context, status, std::forward<decltype(args)>(args)...);
-            },
-            boost::asio::detached);
-        grpc_context.run();
-
-        if (! status.ok()) // TODO: handle different kind of status codes
-        {
-            if (plugin_info_->has_value())
-            {
-                throw exceptions::RemoteException(*slot_info_, plugin_info_->value(), status.error_message());
-            }
-            throw exceptions::RemoteException(*slot_info_, status.error_message());
-        }
-    }
-
-private:
-    template<plugins::v0::SlotID Subscription>
-    boost::asio::awaitable<void> broadcastCall(agrpc::GrpcContext& grpc_context, grpc::Status& status, auto&&... args)
-    {
-        grpc::ClientContext client_context{};
-        prep_client_context(client_context, *slot_info_);
-
-        auto broadcaster{ details::broadcast_factory<broadcast_stub_t, Subscription>() };
-        auto request = details::broadcast_message_factory<Subscription>(std::forward<decltype(args)>(args)...);
-        auto response = google::protobuf::Empty{};
-        status = co_await broadcaster.request(grpc_context, broadcast_stub_, client_context, request, response, boost::asio::use_awaitable);
-        co_return;
-    }
-
-    details::slot_info_ptr slot_info_;
-    details::plugin_info_ptr plugin_info_;
-    ranges::semiregular_box<broadcast_stub_t> broadcast_stub_; ///< The gRPC Broadcast stub for communication.
-};
 
 /**
  * @brief A plugin proxy class template.
